Entregables/Claudia Nohemi Miranda Vildozola/APPIUM/requirements.txt
behave == 1.2.6
Appium-Python-Client==2.9.0
allure-behave==2.13.2
python-dotenv==1.0.0
async-generator==1.10
attrs==23.1.0
certifi==2022.12.7
cffi==1.15.1
colorama==0.4.6
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
iniconfig==2.0.0
outcome==1.2.0
packaging==23.1
parse==1.19.0
parse-type==0.6.0
pluggy==1.0.0
pycparser==2.21
PySocks==1.7.1
pytest==7.3.1
selenium==4.8.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.10.2
urllib3==1.26.15
wsproto==1.2.0

